All changes for new description engine.
git-svn-id: svn+ssh://50.28.74.132/home/luminari/svn/luminari/myproject/trunk@2863 dda7f814-ccd3-4591-a6d7-1bd23e5e64af

+++ mysql.c
@@ -406,13 +406,13 @@ struct region_proximity_list* get_nearby_regions(zone_rnum zone, int x, int y, i
     CREATE(new_node, struct region_proximity_list, 1);
     new_node->rnum = real_region(atoi(row[0]));    
     new_node->n  = atof(row[1]);
-    new_node->ne = atoi(row[2]);    
-    new_node->e  = atoi(row[3]);
-    new_node->se = atoi(row[4]);
-    new_node->s  = atoi(row[5]);
-    new_node->sw = atoi(row[6]);
-    new_node->w  = atoi(row[7]);
-    new_node->nw = atoi(row[8]);
+    new_node->ne = atof(row[2]);    
+    new_node->e  = atof(row[3]);
+    new_node->se = atof(row[4]);
+    new_node->s  = atof(row[5]);
+    new_node->sw = atof(row[6]);
+    new_node->w  = atof(row[7]);
+    new_node->nw = atof(row[8]);
     new_node->dist = atof(row[9]);
     
     new_node->next = regions;
